fix: allow the user to bypass the default IPv6 endpoint
Somehow in the http->start() function, the default behaviour when there is no http->config.address specified is to create an IPv6 endpoint.
There was no way to specify an address in the library for the user so I added one.
The way it has been implemented should be leave the behaviour unchanged for any existing code.